requirements.txt
Flask
Flask-Cors
Flask-GraphQL
graphene
psycopg2-binary
